query operations for all and where

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean
+from sqlalchemy import Column, Float, Integer, String, Boolean
 from sqlalchemy.sql.expression import text
 from sqlalchemy.sql.sqltypes import TIMESTAMP
 
@@ -9,7 +9,7 @@ class Sale(Base):
 
     id = Column(Integer, primary_key=True, nullable=False)
     items = Column(String, nullable=False)
-    total = Column(Integer, nullable=False)
+    total = Column(Float, nullable=False)
     payment_method = Column(String, nullable=False)
     credit = Column(Boolean, server_default='False', nullable=False)
     created_at = Column(TIMESTAMP(timezone=True), nullable=False, server_default=text('now()'))
